requirements.txt
python-telegram-bot
pony

# PycURL